+++ templates/boilerplate/bower-json.js
