+++ templates/boilerplate/gitignore.js
