+++ templates/boilerplate/origami-manifest.js
